fixed some string format types

+++ src/systemcmds/tests/test_automation.cpp
@@ -68,13 +68,13 @@ void vehicle_local_position_uORB_topic::dump()
 DECLARE_UORB_SUB(vehicle_global_position)
 void vehicle_global_position_uORB_topic::dump()
 {
-	PX4_DEBUG("lat = %f, lon = %f, alt =%f, eph = %f, epv = %f, timestamp = %lu",
+	PX4_DEBUG("lat = %f, lon = %f, alt =%f, eph = %f, epv = %f, timestamp = %llu",
 		  vehicle_global_position_status.lat,
 		  vehicle_global_position_status.lon,
 		  vehicle_global_position_status.alt,
 		  vehicle_global_position_status.eph,
 		  vehicle_global_position_status.epv,
-		  vehicle_global_position_status.timestamp);
+		  (long long unsigned) vehicle_global_position_status.timestamp);
 }
 
 DECLARE_UORB_SUB(vehicle_land_detected)
